Moved unit tests related to GetResourceKanjiDecks function into KanjiDeckServiceTests.
