+++ AnkiJapaneseFlashcardManagerTest/Anki2DeckControllerTests.cs
@@ -45,35 +45,6 @@ namespace AnkiJapaneseFlashcardManagerTests
 		}
 
 		[Theory]
-		[InlineData("empty_kanjiResource_deck.anki2")]
-		[InlineData("empty_kanjiResource_newKanji_decks.anki2")]
-		public void Get_kanji_resource_decks(string anki2File)//TODO: Refactor to check for expected values
-		{
-			//Arange
-			Anki2Controller anki2Controller = new Anki2Controller(_anki2FolderPath + anki2File);
-
-			//Act
-			var taggedDecks = anki2Controller.GetResourceKanjiDecks();
-
-			//Assert
-			taggedDecks.Should().NotBeEmpty();
-		}
-
-		[Theory]
-		[InlineData("empty_random_decks.anki2")]
-		public void No_kanji_resource_decks_is_empty(string anki2File)
-		{
-			//Arange
-			Anki2Controller anki2Controller = new Anki2Controller(_anki2FolderPath + anki2File);
-
-			//Act
-			var taggedDecks = anki2Controller.GetResourceKanjiDecks();
-
-			//Assert
-			taggedDecks.Should().BeEmpty();
-		}
-
-		[Theory]
 		[InlineData("empty_newKanji_deck.anki2")]
 		[InlineData("empty_kanjiResource_newKanji_decks.anki2")]
 		public void Get_new_kanji_decks(string anki2File)//TODO: Refactor to check for expected values
